Переписано на Spring Data JPA
