+++ bookshelf/src/main/java/ru/otus/homework/repository/BookRepositoryCustom.java
@@ -1,0 +1,11 @@
+package ru.otus.homework.repository;
+
+import ru.otus.homework.model.Book;
+
+import java.util.Optional;
+
+public interface BookRepositoryCustom{
+
+    Optional<Book> findByIdWithComments(long id);
+
+}
