--- bookshelf/src/main/java/ru/otus/homework/repository/impl/AuthorRepositoryImpl.java
@@ -1,42 +1,0 @@
-package ru.otus.homework.repository.impl;
-
-import org.springframework.stereotype.Repository;
-import org.springframework.transaction.annotation.Transactional;
-import ru.otus.homework.model.Author;
-import ru.otus.homework.repository.AuthorRepository;
-
-import javax.persistence.EntityManager;
-import javax.persistence.PersistenceContext;
-import javax.persistence.TypedQuery;
-import java.util.List;
-import java.util.Optional;
-
-@Repository
-public class AuthorRepositoryImpl implements AuthorRepository {
-
-    @PersistenceContext
-    private EntityManager em;
-
-    @Override
-    @Transactional
-    public Author save(Author author) {
-        if (author.getId() == 0) {
-            em.persist(author);
-            return author;
-        } else {
-            return em.merge(author);
-        }
-    }
-
-    @Override
-    public Optional<Author> findById(long id) {
-        return Optional.ofNullable(em.find(Author.class, id));
-    }
-
-    @Override
-    public List<Author> findAll() {
-        TypedQuery<Author> query = em.createQuery("from Author", Author.class);
-        return query.getResultList();
-    }
-
-}
