--- bookshelf/src/main/java/ru/otus/homework/repository/impl/CommentRepositoryImpl.java
@@ -1,30 +1,0 @@
-package ru.otus.homework.repository.impl;
-
-import org.springframework.stereotype.Repository;
-import ru.otus.homework.model.Comment;
-import ru.otus.homework.repository.CommentRepository;
-
-import javax.persistence.EntityManager;
-import javax.persistence.PersistenceContext;
-import javax.persistence.TypedQuery;
-import java.util.List;
-import java.util.Optional;
-
-@Repository
-public class CommentRepositoryImpl implements CommentRepository {
-
-    @PersistenceContext
-    private EntityManager em;
-
-    @Override
-    public Optional<Comment> findById(long id) {
-        return Optional.ofNullable(em.find(Comment.class, id));
-    }
-
-    @Override
-    public List<Comment> findAll() {
-        TypedQuery<Comment> query = em.createQuery("from Comment", Comment.class);
-        return query.getResultList();
-    }
-
-}
