--- bookshelf/src/main/java/ru/otus/homework/repository/impl/GenreRepositoryImpl.java
@@ -1,41 +1,0 @@
-package ru.otus.homework.repository.impl;
-
-import org.springframework.stereotype.Repository;
-import org.springframework.transaction.annotation.Transactional;
-import ru.otus.homework.model.Genre;
-import ru.otus.homework.repository.GenreRepository;
-
-import javax.persistence.EntityManager;
-import javax.persistence.PersistenceContext;
-import javax.persistence.TypedQuery;
-import java.util.List;
-import java.util.Optional;
-
-@Repository
-public class GenreRepositoryImpl implements GenreRepository {
-
-    @PersistenceContext
-    private EntityManager em;
-
-    @Override
-    @Transactional
-    public Genre save(Genre genre) {
-        if (genre.getId() == 0) {
-            em.persist(genre);
-            return genre;
-        } else {
-            return em.merge(genre);
-        }
-    }
-
-    @Override
-    public Optional<Genre> findById(long id) {
-        return Optional.ofNullable(em.find(Genre.class, id));
-    }
-
-    @Override
-    public List<Genre> findAll() {
-        TypedQuery<Genre> query = em.createQuery("from Genre", Genre.class);
-        return query.getResultList();
-    }
-}
