+++ bookshelf/src/test/java/ru/otus/homework/repository/impl/BookRepositoryCustomImplTest.java
@@ -10,58 +10,22 @@ import ru.otus.homework.model.Author;
 import ru.otus.homework.model.Book;
 import ru.otus.homework.model.Comment;
 import ru.otus.homework.model.Genre;
+import ru.otus.homework.repository.BookRepositoryCustom;
 
-import javax.persistence.TypedQuery;
-import java.util.List;
 import java.util.Set;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 @DisplayName("Репозиторий книг должен:")
 @DataJpaTest
-@Import(BookRepositoryImpl.class)
-class BookRepositoryImplTest {
+@Import(BookRepositoryCustomImpl.class)
+class BookRepositoryCustomImplTest {
 
     @Autowired
-    private BookRepositoryImpl repository;
+    private BookRepositoryCustom repository;
 
     @Autowired
     private TestEntityManager em;
-
-    @DisplayName("возвращать книгу после сохранения;")
-    @Test
-    void shouldReturnBookAfterSave() {
-        Book newBook = Book.builder()
-                .name("Test Book")
-                .build();
-        assertThat(repository.save(newBook)).extracting("name").isEqualTo("Test Book");
-    }
-
-    @DisplayName("возвращать книгу по ID;")
-    @Test
-    void shouldReturnBookById() {
-        Book newBook = Book.builder().name("Test Book").build();
-        Book persistedBook = em.persist(newBook);
-        assertThat(repository.findById(persistedBook.getId()).get()).isEqualTo(persistedBook);
-    }
-
-    @DisplayName("возвращать список всех книг;")
-    @Test
-    void shouldReturnAllBooks() {
-        TypedQuery<Book> query = em.getEntityManager().createQuery("from Book", Book.class);
-        List<Book> actualBooks = query.getResultList();
-        assertThat(repository.findAll()).containsExactlyElementsOf(actualBooks);
-    }
-
-    @DisplayName("корректно удалять книгу;")
-    @Test
-    void shouldDelete() {
-        Book newBook = Book.builder().name("Test Book").build();
-        Book persistedBook = em.persist(newBook);
-        long persistedBookId = persistedBook.getId();
-        repository.delete(persistedBook);
-        assertThat(em.find(Book.class, persistedBookId)).isNull();
-    }
 
     @DisplayName("возвращать книгу по ID с комментариями;")
     @Test
